Simplify shell-websocket.service and add authToken to connection flow
+ Removes base-shell-websocket.service and ssh-shell-websocket.service
now that ssh and legacy ssm connections no longer need to be supported

+ Adds connectionNodeId and authToken to constructor and use when
creating the signalR hub connection
