+++ shell-websocket.service/shell-websocket.service.types.ts
@@ -4,9 +4,9 @@ import { IDisposableAsync } from '../utility/disposable';
 export interface IShellWebsocketService extends IDisposableAsync{
     start() : Promise<void>;
 
-    // Terminal dimensions provided in the shell connect are only used by ssh
-    // and non-keysplitting ssm shells where terminal size is set from the
-    // backend. Keysplitting ssm shells must send a separate resize input event
+    // Terminal dimensions provided in the shell connect are only used for the
+    // virtual terminal emulator for command extraction. Keysplitting shells
+    // must send a separate resize input event
     sendShellConnect(rows: number, cols: number, replayOutput: boolean): void;
 
     sendReplayDone(rows: number, cols: number): void;
